Update project for Spark 3.5.0
+ Separate examples to subproject
