+++ project/plugins.sbt
@@ -1,11 +1,1 @@
-resolvers += "Spark Package Main Repo" at "https://dl.bintray.com/spark-packages/maven"
-
-addSbtPlugin("io.spray" % "sbt-revolver" % "0.8.0")
-
-addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "0.6.0")
-
-addSbtPlugin("org.spark-packages" % "sbt-spark-package" % "0.2.1")
-
-addSbtPlugin("com.typesafe.sbt" % "sbt-scalariform" % "1.3.0")
-
-
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "2.2.0")
